+++ public/js/patron_trails.js
@@ -63,22 +63,28 @@ addPatronTrailForm.addEventListener("submit", function (e) {
 });
 
 function deletePatronTrail(patronName, trailName) {
-  let link = "/patron_trails";
-  let data = {
-    patron_name: patronName,
-    trail_name: trailName,
-  };
+  if (
+    window.confirm(
+      "Do you really want to delete this record? This cannot be undone. Click 'OK' to continue with deleting this record."
+    )
+  ) {
+    let link = "/patron_trails";
+    let data = {
+      patron_name: patronName,
+      trail_name: trailName,
+    };
 
-  console.log(`data before: ${JSON.stringify(data)}`);
+    console.log(`data before: ${JSON.stringify(data)}`);
 
-  $.ajax({
-    url: link,
-    type: "DELETE",
-    data: JSON.stringify(data),
-    contentType: "application/json; charset=utf-8",
-    success: function (result) {
-      // Refresh page to requery updated table
-      location.reload();
-    },
-  });
+    $.ajax({
+      url: link,
+      type: "DELETE",
+      data: JSON.stringify(data),
+      contentType: "application/json; charset=utf-8",
+      success: function (result) {
+        // Refresh page to requery updated table
+        location.reload();
+      },
+    });
+  }
 }
