+++ public/js/patrons.js
@@ -179,24 +179,30 @@ function savePatronEdit(patronID) {
 }
 
 function deletePatron(patronID) {
-  let link = "/patrons";
-  let data = {
-    id: patronID,
-  };
+  if (
+    window.confirm(
+      "Do you really want to delete this record? This cannot be undone. Click 'OK' to continue with deleting this record."
+    )
+  ) {
+    let link = "/patrons";
+    let data = {
+      id: patronID,
+    };
 
-  console.log(`data before: ${JSON.stringify(data)}`);
+    console.log(`data before: ${JSON.stringify(data)}`);
 
-  $.ajax({
-    url: link,
-    type: "DELETE",
-    data: JSON.stringify(data),
-    contentType: "application/json; charset=utf-8",
-    success: function (result) {
-      // Refresh page to requery updated table
-      location.reload();
-      // deleteRow(patronID);
-    },
-  });
+    $.ajax({
+      url: link,
+      type: "DELETE",
+      data: JSON.stringify(data),
+      contentType: "application/json; charset=utf-8",
+      success: function (result) {
+        // Refresh page to requery updated table
+        location.reload();
+        // deleteRow(patronID);
+      },
+    });
+  }
 }
 
 // function deleteRow(patronID) {
