+++ public/js/rewards_points.js
@@ -46,21 +46,27 @@ addRewardForm.addEventListener("submit", function (e) {
 });
 
 function deleteReward(rewardID) {
-  let link = "/rewards_points";
-  let data = {
-    id: rewardID,
-  };
+  if (
+    window.confirm(
+      "Do you really want to delete this record? This cannot be undone. Click 'OK' to continue with deleting this record."
+    )
+  ) {
+    let link = "/rewards_points";
+    let data = {
+      id: rewardID,
+    };
 
-  console.log(`data before: ${JSON.stringify(data)}`);
+    console.log(`data before: ${JSON.stringify(data)}`);
 
-  $.ajax({
-    url: link,
-    type: "DELETE",
-    data: JSON.stringify(data),
-    contentType: "application/json; charset=utf-8",
-    success: function (result) {
-      // Refresh page to requery updated table
-      location.reload();
-    },
-  });
+    $.ajax({
+      url: link,
+      type: "DELETE",
+      data: JSON.stringify(data),
+      contentType: "application/json; charset=utf-8",
+      success: function (result) {
+        // Refresh page to requery updated table
+        location.reload();
+      },
+    });
+  }
 }
